new static chart legend
