+++ frontend/src/metabase/static-viz/components/Legend/constants.ts
@@ -1,3 +1,3 @@
 export const LEGEND_CIRCLE_SIZE = 10;
-export const LEGEND_CIRCLE_MARGIN_RIGHT = 16;
+export const LEGEND_CIRCLE_MARGIN_RIGHT = 8;
 export const LEGEND_ITEM_MARGIN_RIGHT = 16;
