+++ frontend/src/metabase/static-viz/components/Legend/index.ts
