+++ frontend/src/metabase/static-viz/components/Legend/types.ts
